Added editing of connection profiles to identities
Updated login screen to edit connection profiles

contributes to hyperledger/composer#1425
