+++ packages/composer-playground/src/app/app.module.ts
@@ -20,7 +20,6 @@ import { AppState, InternalStateType } from './app.service';
 import { AboutComponent } from './about';
 import { BasicModalsModule } from './basic-modals/basic-models.module';
 import { WelcomeComponent } from './welcome';
-import { LoginComponent } from './login';
 import { NoContentComponent } from './no-content';
 import { VersionCheckComponent } from './version-check';
 import { FooterComponent } from './footer';
@@ -59,7 +58,6 @@ type StoreType = {
     declarations: [
         AboutComponent,
         AppComponent,
-        LoginComponent,
         NoContentComponent,
         VersionCheckComponent,
         WelcomeComponent
