--- packages/composer-playground/src/app/connection-profile/connection-profile-data/index.ts
@@ -1,1 +1,0 @@
-export * from './connection-profile-data.component';
