+++ packages/composer-playground/src/app/connection-profile/connection-profile.component.scss
@@ -2,72 +2,228 @@
 @import '../../assets/styles/base/_variables.scss';
 
 connection-profile {
-  display: flex;
-  width: 100%;
 
-  .side-bar {
-    display: flex;
-    flex-direction: column;
+    $highlighted: #e4e4e4;
 
-    .profiles {
-      & > * {
-        padding: $space-medium $space-large;
-      }
-      margin-top: $space-large;
-
-      span {
-        color: $secondary-text;
-      }
-
-      .side-bar-nav {
-        padding: 0;
-      }
-
-    }
-
-    .about {
-      flex: 1;
-      justify-content: flex-end;
-      padding: $space-medium $space-large;
-    }
-  }
-
-  .main-view {
-    overflow-y: auto;
-    background-color: $fourth-highlight;
-    padding: $space-large $space-large $space-medium $space-large;
-    display: flex;
-    flex-direction: column;
-
-    .playground-profile-warning {
-      background-color: $white;
-      border: 1px solid $fourth-highlight;
-      border-radius: 0.25em;
-      padding: $space-medium;
-      margin-bottom: $space-large;
-
-      .profile-warning-content {
+    .profile-header {
         display: flex;
+        color: $primary-text;
+        flex-direction: column;
 
-        .profile-warning-text {
-          margin: 0 $space-smedium;
+        p {
+            padding-bottom: $space-smedium;
         }
-      }
+
+        .profile-header-details {
+            flex: 1;
+            display: flex;
+            align-items: baseline;
+
+            .profile-title {
+                flex: 1;
+            }
+            .profile-actions {
+                flex-shrink: 1;
+            }
+        }
+
     }
 
-    .business-network-details {
-      display: flex;
-
-      .business-network-version {
-        margin-right: $space-smedium;
-        color: $secondary-text;
-      }
-      .margin-right {
-        margin-right: $space-smedium;
-      }
-      .edit-package-text {
-        font-style: italic;
-      }
+    .profile-data-container {
+        background-color: $white;
+        border-radius: 3px;
+        padding: $space-smedium;
+        display: flex;
+        flex-flow: column;
+        margin-bottom: $space-medium;
     }
-  }
+
+    .data-section {
+        max-height: 0px;
+        font-size: 14px;
+        @include transition(all);
+        overflow: hidden;
+
+        &.expanded {
+            @include transition(all);
+            max-height: 3000px; // Need this to get transition to work
+            overflow: hidden;
+            padding-bottom: 0px;
+        }
+    }
+
+    .profile-data-header {
+        flex: 1;
+        display: flex;
+        flex-flow: column;
+        padding: 0 $space-medium;
+
+        .section-details {
+            display: flex;
+            flex-direction: row;
+            flex: 1;
+            border-bottom: solid 1px #D0DADA;
+            margin-bottom: $space-medium;
+            align-items: center;
+
+            .section-name {
+                flex: 1;
+                padding: $space-medium 0 $space-medium;
+                cursor: pointer;
+                font-family: $font-stack;
+                font-weight: 600;
+                font-size: 14px;
+                color: #5068c2;
+                letter-spacing: 0;
+                text-align: left;
+            }
+            .section-expand {
+                button {
+                    line-height: 30px;
+                    font-size: 0.8rem;
+                }
+            }
+        }
+
+    }
+    .profile-row {
+        flex: 2;
+        display: flex;
+        flex-flow: row;
+        color: $primary-text;
+        font-size: 14px;
+        padding-bottom: $space-medium;
+
+        .profile-property {
+            align-self: flex-start;
+            flex-basis: 20%;
+            padding-right: $space-large;
+            font-size: 14px;
+            font-weight: 600;
+        }
+
+        .profile-value {
+            flex-basis: 50%;
+            display: flex;
+            flex-direction: column;
+            .nested-group {
+                display: flex;
+                flex-direction: column;
+            }
+            .nested-section {
+                border-left: 4px solid $first-highlight;
+                padding-left: $space-smedium;
+                display: flex;
+                flex-direction: column;
+                margin-bottom: $space-medium;
+                .nested-row {
+                    flex: 1;
+                    display: flex;
+                    align-items: center;
+                    padding-bottom: $space-medium;
+                    .nested-property {
+                        flex-basis: 20%;
+                        padding: 0 $space-smedium $space-small 0;
+                        font-weight: 600;
+                    }
+                    .nested-value {
+                        flex: 1;
+                        display: flex;
+                        flex-direction: column;
+
+                    }
+                }
+                .nested-row:last-child {
+                    padding-bottom: 0;
+                }
+            }
+
+        }
+
+        .profile-buttons {
+            align-self: flex-end;
+            padding-bottom: $space-medium;
+        }
+
+    }
+    .profile-row:last-child {
+        margin-bottom: 0;
+    }
+
+    .chevron {
+        @include transition(all);
+        transform: rotate(-90deg);
+    }
+
+    .edit-data {
+        background-color: $white;
+        margin-top: $space-medium;
+
+        section {
+            padding: $space-medium;
+            margin-bottom: $space-large;
+
+            .section-title {
+                border-bottom: 1px solid $fourth-highlight;
+                padding-bottom: $space-medium;
+                margin-bottom: $space-medium;
+                font-size: 14px;
+                font-weight: 600;
+                color: $first-highlight;
+            }
+
+            div {
+                display: flex;
+
+                label {
+                    flex-basis: 20%;
+                }
+
+            }
+
+            .actions {
+                display: flex;
+                background-color: $third-highlight;
+                padding: $space-medium;
+
+                .info {
+                    flex: 1
+                }
+            }
+        }
+    }
+
+    .orderer-group {
+        display: flex;
+        flex-direction: column;
+        .orderer-content {
+            flex: 1;
+            display: flex;
+            flex-direction: column;
+            .orderer-entry {
+                flex: 1;
+            }
+        }
+    }
+
+    .certificate-button {
+        @extend button;
+        padding-right: 10px;
+        padding-left: 10px;
+        min-width: 0;
+        &:hover {
+            background-color: $white;
+            border-color: $second-highlight;
+            svg {
+                fill: #3F55AF
+            }
+
+        }
+    }
+
+    .no-cert {
+        border-color: $error-colour-1;
+        background-color: $error-colour-2;
+    }
+
 }
