--- packages/composer-playground/src/app/connection-profile/delete-connection-profile/delete-connection-profile.component.html
@@ -1,25 +1,0 @@
-<div class="delete-connection-profile">
-  <div class="modal-header">
-    <svg class="ibm-icon" aria-hidden="true">
-      <use xlink:href="#icon-warn_32"></use>
-    </svg>
-    <h1>Delete Connection Profile</h1>
-    <button class="icon modal-exit" (click)="activeModal.dismiss()">
-      <svg class="ibm-icon" aria-hidden="true">
-        <use xlink:href="#icon-close_new"></use>
-      </svg>
-    </button>
-  </div>
-  <section class="modal-body">
-    <span>Are you sure you want to delete this connection profile?</span>
-
-  </section>
-  <footer>
-    <button type="button" class="secondary" (click)="activeModal.dismiss();">
-      <span>Cancel</span>
-    </button>
-    <button type="button" class="primary" (click)="deleteProfile();">
-      <span>Delete</span>
-    </button>
-  </footer>
-</div>
