--- packages/composer-playground/src/app/connection-profile/delete-connection-profile/delete-connection-profile.component.scss
@@ -1,26 +1,0 @@
-@import '../../../assets/styles/base/colors';
-@import '../../../assets/styles/base/variables';
-
-.delete-connection-profile {
-  width: 660px;
-  height: 200px;
-
-  display: flex;
-  flex-direction: column;
-
-  .modal-header {
-    svg {
-      fill: $error-colour-1
-    }
-  }
-
-  section {
-    font-size: 0.9em;
-
-    span {
-      color: $secondary-text;
-    }
-
-  }
-}
-
