--- packages/composer-playground/src/app/connection-profile/delete-connection-profile/index.ts
@@ -1,1 +1,0 @@
-export * from './delete-connection-profile.component';
