+++ packages/composer-playground/src/app/login/login-routing.module.ts
@@ -1,12 +1,12 @@
 import { NgModule }            from '@angular/core';
 import { RouterModule }        from '@angular/router';
 
-import { ConnectionProfileComponent }    from './connection-profile.component';
+import { LoginComponent }    from './login.component';
 
 @NgModule({
     imports: [RouterModule.forChild([
-        { path: '', component: ConnectionProfileComponent }
+        { path: '', component: LoginComponent }
     ])],
     exports: [RouterModule]
 })
-export class ConnectionProfileRoutingModule {}
+export class LoginRoutingModule {}
