+++ packages/composer-playground/src/app/login/login.component.ts
@@ -5,9 +5,10 @@ import { AdminService } from '../services/admin.service';
 import { ConnectionProfileService } from '../services/connectionprofile.service';
 import { ClientService } from '../services/client.service';
 import { InitializationService } from '../services/initialization.service';
+import { AlertService } from '../basic-modals/alert.service';
 
 @Component({
-    selector: 'login',
+    selector: 'app-login',
     templateUrl: './login.component.html',
     styleUrls: [
         './login.component.scss'.toString()
@@ -15,44 +16,59 @@ import { InitializationService } from '../services/initialization.service';
 })
 export class LoginComponent implements OnInit {
 
-    private identities = {};
     private connectionProfiles = [];
+    private editingConectionProfile = null;
 
     constructor(private identityService: IdentityService,
                 private router: Router,
                 private adminService: AdminService,
                 private connectionProfileService: ConnectionProfileService,
                 private clientService: ClientService,
-                private initializationService: InitializationService) {
+                private initializationService: InitializationService,
+                private alertService: AlertService) {
 
     }
 
     ngOnInit() {
         return this.initializationService.initialize()
             .then(() => {
-                return this.connectionProfileService.getAllProfiles();
-            })
-            .then((profiles) => {
-                this.connectionProfiles = Object.keys(profiles);
-                this.connectionProfiles.forEach((profile) => {
-                    this.identities[profile] = [];
-                    return this.identityService.getIdentities(profile)
-                        .then((identities) => {
-                            identities.forEach((identity) => {
-                                this.identities[profile].push({
-                                    userId: identity,
-                                    connectionProfile: profile,
-                                    businessNetwork: 'org-acme-biznet'
-                                });
-                            });
-                        });
-                });
+                return this.loadConnectionProfiles();
             });
     }
 
-    changeIdentity(identity): Promise<boolean> {
-        this.connectionProfileService.setCurrentConnectionProfile(identity.connectionProfile);
-        this.identityService.setCurrentIdentity(identity.userId);
+    loadConnectionProfiles(): Promise<void> {
+        return this.connectionProfileService.getAllProfiles()
+            .then((profiles) => {
+                let newConnectionProfiles = [];
+                let keys = Object.keys(profiles).sort();
+                keys.forEach((key) => {
+                    return this.identityService.getIdentities(key)
+                        .then((identities) => {
+                            let identityList = [];
+                            identities.forEach((identity) => {
+                                identityList.push({
+                                    userId: identity,
+                                    businessNetwork: 'org-acme-biznet'
+                                });
+                            });
+
+                            let connectionProfile = profiles[key];
+                            newConnectionProfiles.push({
+                                name: key,
+                                profile: connectionProfile,
+                                default: key === '$default',
+                                identities: identityList
+                        });
+                });
+            });
+
+                this.connectionProfiles = newConnectionProfiles;
+            });
+    }
+
+    changeIdentity(connectionProfile, userId): Promise<boolean | void> {
+        this.connectionProfileService.setCurrentConnectionProfile(connectionProfile);
+        this.identityService.setCurrentIdentity(userId);
         return this.adminService.list()
             .then((businessNetworks) => {
                 return this.clientService.ensureConnected(businessNetworks[0], true);
@@ -60,6 +76,19 @@ export class LoginComponent implements OnInit {
             .then(() => {
                 this.identityService.setLoggedIn(true);
                 return this.router.navigate(['editor']);
+            })
+            .catch((error) => {
+                this.alertService.errorStatus$.next(error);
             });
+
+    }
+
+    editConnectionProfile(connectionProfile): void {
+        this.editingConectionProfile = connectionProfile;
+    }
+
+    finishedEditingConnectionProfile(): Promise<void> {
+        delete this.editingConectionProfile;
+        return this.loadConnectionProfiles();
     }
 }
